src/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-appconf==1.0.3
django-cleanup==4.0.0
django-imagekit==4.0.2
image==1.5.27
pilkit==2.0
Pillow==7.0.0
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
